Module Tracelog.MakeDebug

This is useful when developping and debugging a specific module: we can enable debugging forthis module by replacing the Make call by MakeDebug.

Parameters

module Category : sig ... end

Signature

val error : 'a log -> unit

Printing from most to least important. info and debug levels are not printed by default.

val warning : 'a log -> unit

val notice : 'a log -> unit

val info : 'a log -> unit

val debug : 'a log -> unit

val fatal : 'a log -> 'b

Classic error messages. Fatal will raise the Fatal exception.

val not_yet_implemented : 'a log -> unit

val trace : 'a log -> ?loc:location -> ?bintrace:'c Syntax_tree.Location_identifier.t -> ?pp_ret:(Format.formatter -> 'b -> unit) -> (unit -> 'b) -> 'b

trace log pp_ret f first display log, then executes f and prints its return value using pp_ret.

val fatal_handler : ('a -> 'b) -> 'a -> 'b

fatal_handler Fatal handler, will properly collect the backtrace for fatal and print it. It also collects other uncaught exception and returns an error message.